+++ clicks/getclickdata.go
@@ -28,7 +28,7 @@ func errorGet(c *gin.Context, err error, reason string) {
 	})
 }
 
-func GetClicksDB(db *gorm.DB, paramkey int, user string) (datatypes.Entry, []datatypes.Click, error, bool) {
+func GetClicksDB(db *gorm.DB, paramkey int64, user string) (datatypes.Entry, []datatypes.Click, error, bool) {
 	var entry datatypes.Entry
 
 	err := db.Where("id = ? AND user = ? AND archived = ?", paramkey, user, false).First(&entry).Error
